requirements.txt
boto3>=1.34.0
pyyaml>=6.0
click>=8.1.0
pylint>=3.0.0
pytest>=7.4.0
pre-commit>=3.5.0
